+++ src/main/groovy/cz/bh/lisp/parser/sexp/IntegerNode.groovy
@@ -1,7 +1,7 @@
 package cz.bh.lisp.parser.sexp
 
-class IntegerNode extends Node<Integer> {
-    IntegerNode(Integer val, int line) {
+class IntegerNode extends Node<BigInteger> {
+    IntegerNode(BigInteger val, int line) {
         super(val, line)
     }
 }
